Introduced the Acts::Cuda::ResultScalar type.
It is a replacement for the Acts::CudaScalar and Acts::CpuScalar types.
Unlike those two, it is *only* meant to be used to pass scalar
result variables from a CUDA kernel back to the host.

At the same time removed all the (now) unnecessary classes, and re-designed
how scalar function arguments would be passed to the CUDA kernels.