Introduced a first step for the triplet finding.
At the same time got rid of the Acts::Cuda::DeviceMatrix type. Not only was
it over-complicated, but it turned out to be very slow as well. Instead of that
fancy code, I now introduced a simple macro for accessing a single element of
a 2-dimensional matrix (represented as a 1-dimensional array in memory).