Made the duplet search fill a vecmem::jagged_vector with its output.
And then the code still converts it back into "old style" arrays
in memory, so that the later steps would not have had to be
converted already.

The implementation works (even if slowly) on the host and on the iGPU.
But it fails with the CPU device. Possibly due to some issue in the CPU
driver... (Would not be the first time for jagged vectors not to work
correctly with the OpenCL CPU driver. :-( )